refactor: use assertFactory in tests
