+++ test/nu/npm.spec.ts
@@ -1,13 +1,8 @@
-import { expect, test } from 'vitest'
-import { parseNu } from '../../src/commands'
+import { test } from 'vitest'
+import { parseNu } from '../../src'
+import { assertFactory } from '../assert'
 
 const agent = 'npm'
-const _ = (arg: string, expected: string) => () => {
-  expect(
-    parseNu(agent, arg.split(' ').filter(Boolean)),
-  ).toBe(
-    expected,
-  )
-}
+const _ = assertFactory(parseNu, agent)
 
 test('empty', _('', 'npm update'))
